OF-3084 OF-3088: Refactor Cache Locking implementation
Based on the single responsibility principle, the CacheFactory should not be managing Locks. Instead, Caches should do that themselves.

Back in 2019, the method in CacheFactory that generates locks was marked as deprecated because of that, with a removal note in the distant future: 5.0.0.

As 5.0.0 is upon us, the deprecated code has now been removed. The cache locking implementation has been moved to a dedicated class (`LocalCache`) that is used by caches directly.

This commit attempts to achieve the desired removal of deprecated methods with a minimum of refactoring.

Further improvements may be desired (e.g. not storing locks for all caches in one shared data structure, but rather in a per-cache data structure).
