+++ xmppserver/src/main/java/org/jivesoftware/util/cache/CacheFactory.java
@@ -604,31 +604,6 @@ public class CacheFactory {
         }
     }
 
-    /**
-     * @deprecated in favour of {@link Cache#getLock}. Will be removed in Openfire 5.0.0.
-     *
-     * <p>Returns an existing {@link java.util.concurrent.locks.Lock} on the specified key or creates a new one
-     * if none was found. This operation is thread safe. Successive calls with the same key may or may not
-     * return the same {@link java.util.concurrent.locks.Lock}. However, different threads asking for the
-     * same Lock at the same time will get the same Lock object.<p>
-     *
-     * The supplied cache may or may not be used depending whether the server is running on cluster mode
-     * or not. When not running as part of a cluster then the lock will be unrelated to the cache and will
-     * only be visible in this JVM.
-     *
-     * @param key the object that defines the visibility or scope of the lock.
-     * @param cache the cache used for holding the lock.
-     * @return an existing lock on the specified key or creates a new one if none was found.
-     */
-    @Deprecated(since = "4.5", forRemoval = true)
-    public static synchronized Lock getLock(Object key, Cache cache) {
-        if (localOnly.contains(cache.getName())) {
-            return localCacheFactoryStrategy.getLock(key, cache);
-        } else {
-            return cacheFactoryStrategy.getLock(key, cache);
-        }
-    }
-
     @SuppressWarnings("unchecked")
     private static <T extends Cache> T wrapCache(T cache, String name) {
         if ("Routing Components Cache".equals(name)) {
