+++ xmppserver/src/main/java/org/jivesoftware/util/cache/lock/LockAndCount.java
@@ -1,0 +1,14 @@
+package org.jivesoftware.util.cache.lock;
+
+import java.util.concurrent.locks.ReentrantLock;
+
+class LockAndCount
+{
+    final ReentrantLock lock;
+    int count;
+
+    LockAndCount(ReentrantLock lock)
+    {
+        this.lock = lock;
+    }
+}
